+++ src/main/Java/com/gzw/servlet/user/UserServlet.java
@@ -14,7 +14,7 @@ import java.io.IOException;
 
 public class UserServlet extends HttpServlet {
     @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) {
         String method = req.getParameter("method");
         if (method.equals("savepwd")&&method!=null){
             this.updatePwd(req,resp);
@@ -23,7 +23,7 @@ public class UserServlet extends HttpServlet {
     }
 
     @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException { doGet(req, resp);
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) { doGet(req, resp);
     }
